added test batch for quick entry testing
This setup should work for now. This is a test file with 2 batches of 5 records each created from the omoccurences harvard dwc file sent. The images used are dummy flower/plant images because the original image links either don’t work or I wasn’t able to locate the correct ones as of now..